feat: Add search functionality and track search history for unauthenticated users
